Fixed build error of httpx tool
Some tools (e.g. httpx) require gologger in version 1.0.1 and use the
the function gologger.Errorf, but in the Dockerfile the "-u" flag
instructs get to use the network to update the named packages and
their dependencies. Thus, gologger is updated to v1.1.2 where the
function gologger.Errorf seems to be removed.

The removal of the "-u" flag fixes the build error.