Replace CVE-2024-41661 with CVE-2023-50094
On 29 August 2024, MITRE informed GitHub that CVE-2024-41661, which was issued August 2024 is a duplicate of CVE-2023-50094, which was issued December 2023 or January 2024. We rejected CVE-2024-41661 as a duplicate CVE because CVE-2023-50094 was published first and suggest that the maintainers of reNgine replace all instances of CVE-2024-41661 with CVE-2023-50094, including in the changelog and repository security advisory.