Information exposure through an exception
To fix the problem, we need to ensure that detailed exception information is not exposed to the end user. Instead, we should log the exception details on the server and return a generic error message to the user. This can be achieved by modifying the exception handling block to log the error and return a generic response.

- Modify the exception handling block in the `CMSDetector` class to log the exception using the `logger` and return a generic error message.
- Ensure that the logging captures the full stack trace for debugging purposes.